@@ -1,4 +1,4 @@
-package ru.kata.spring.boot_security.demo;
+package com.SpringBoot_security;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
